+++ src/tko/game/quest_config.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 
 class QuestGame:
     def __init__(self):
-        self.goal_xp: int = 0
+        self.goal_xp: float = 0.0
         self.skill: str | None = None
         self.languages: set[str] = set()
         self.active: bool = True
